backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.16,<4
djangorestframework-simplejwt>=5.5,<6
drf-spectacular>=0.29,<1
django-cors-headers>=4.9,<5
psycopg2-binary>=2.9,<3
